Crypto API Python Project
An API was used to see the accurate value of a variation of different cryptocurrencies. This data was then saved to a csv file which the user would be able to access. The data was sorted into a table and cleaned slightly to make it more readable. The data was then visualised in a graph.